[feat] OpenApiController 관련 DTO 그루핑
[feat] OpenApiController 관련 DTO 그루핑
