--- src/main/kotlin/com/chatandpay/api/dto/InquiryRealNameDTO.kt
@@ -1,19 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.common.ApiResponse
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class InquiryRealNameDTO(
-
-
-    var bankTranId: String,
-    val bankCodeStd: String,
-    val accountNum: String,
-    var accountHolderInfoType: String = " ",
-    var accountHolderInfo: String,
-    var tranDtime: String,
-    var payUserId: Long
-
-) : ApiResponse()
