--- src/main/kotlin/com/chatandpay/api/dto/OpenApiDepositWalletDTO.kt
@@ -1,8 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.domain.PayUser
-data class OpenApiDepositWalletDTO (
-    val depositMoney: Int,
-    val accountId: Long,
-    val payUser: PayUser,
-)
