+++ src/main/kotlin/com/chatandpay/api/exception/WalletChargeAttemptException.kt
@@ -1,6 +1,6 @@
 package com.chatandpay.api.exception
 
-import com.chatandpay.api.dto.OpenApiDepositWalletDTO
+import com.chatandpay.api.dto.OpenApiDTO
 import lombok.Getter
 import lombok.RequiredArgsConstructor
 
@@ -8,5 +8,5 @@ import lombok.RequiredArgsConstructor
 @RequiredArgsConstructor
 class WalletChargeAttemptException(
     override val message: String,
-    val openApiDto: OpenApiDepositWalletDTO
+    val openApiDto: OpenApiDTO.OpenApiDepositWalletDTO
 ) : Exception()
